refactor(attestation): groupe instructeur add signature shared
